fix disabled styles
